Python files
Files with the programmed data that I've been mostly trying out what works best and what doesn't. 
My main focus is to create a easy-to-everyone visualization. I chose just to create images of each day so those images could compose a short gif that can be reproduced without the need of html and it would be easier to spread awareness.